requirements.txt
requests>=2.31.0
fal-client>=0.5.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pywebview>=4.4.0
python-multipart>=0.0.9
Pillow>=10.0.0
moviepy>=1.0.3
numpy>=1.26.0
